Convert rebuild tests to Docker Compose tests
Rebuild test cases in nexus_rebuild.rs have been converted into Docker
Compose tests in rebuild.rs. The reason for this change is to make the
tests more representative of real world applications. As an example, the
new tests share replicas over NVMf.

One test (rebuild_lookup) remains in nexus_rebuild.rs. This test cannot
be converted to a Compose test as the functionality is not exposed over
RPC. However, the test is still useful and hence remains.