nexus: log error messages in reconfigure context
When we reconfigure channels, we unwrap() as we do not allow/expect
that to fail in anyway. Although this is still correct, it can actually
failed with for various reasons. We can not determine the  actual error
so we have to account for this by simply continuing and not unwrapping.