jenkins worker is running out of space
Jenkins worker was running out of space because docker images were not
pruned. I decided to put all space reclaiming into a single script that
is run before each e2e stage. So nix GC is no longer run as a service
on worker nodes. The advantage is that we can synchronize GC with running
jobs that way.